+++ labs/memory_bound/data_packing/solution.h
@@ -8,10 +8,10 @@ constexpr int maxRandom = 100;
 
 // FIXME: this data structure can be reduced in size
 struct S {
-  int i;
-  long long l;
+  short i;
+  int l;
   short s;
-  double d;
+  float d;
   bool b;
 
   bool operator<(const S &s) const { return this->i < s.i; }
